dcp-815 separate input.py
