+++ exporter/graph/crawler.py
@@ -4,8 +4,9 @@ from functools import reduce
 from operator import iconcat
 from typing import List, Iterable, Optional, Callable
 
-from exporter.graph.experiment_graph import ExperimentGraph, ProcessLink, Input, Output, \
+from exporter.graph.experiment_graph import ExperimentGraph, ProcessLink, Output, \
     SupplementaryFileLink, SupplementedEntity, SupplementaryFile
+from exporter.graph.input import Input
 from exporter.graph.protocol_link import ProtocolLink
 from exporter.metadata.resource import MetadataResource
 from exporter.metadata.service import MetadataService
